+++ torch_frame/nn/encoder/stype_encoder.py
@@ -3,11 +3,18 @@ from typing import Any, Dict, List, Optional, Set
 
 import torch
 from torch import Tensor
-from torch.nn import Embedding, ModuleList, Parameter
+from torch.nn import Embedding, ModuleList, Parameter, Sequential
 
 from torch_frame import stype
 from torch_frame.data.stats import StatType
 from torch_frame.nn.base import Module
+
+
+def reset_parameters_soft(module: Module):
+    r"""Call reset_parameters() only when it exists. Skip activation module."""
+    if (hasattr(module, 'reset_parameters')
+            and callable(module.reset_parameters)):
+        module.reset_parameters()
 
 
 class StypeEncoder(Module, ABC):
@@ -19,6 +26,10 @@ class StypeEncoder(Module, ABC):
         out_channels (int): The output channel dimensionality
         stats_list (List[Dict[StatType, Any]]): The list of stats for each
             column within the same stype.
+        post_module (Module, optional): The posthoc module applied to the
+            output, such as activation function and normalization. Must
+            preserve the shape of the output. If :obj:`None`, no module will be
+            applied to the output. (default: :obj:`None`)
     """
     supported_stypes: Set[stype] = {}
     LAZY_ATTRS = {'out_channels', 'stats_list'}
@@ -28,8 +39,9 @@ class StypeEncoder(Module, ABC):
         self,
         out_channels: Optional[int] = None,
         stats_list: Optional[List[Dict[StatType, Any]]] = None,
+        post_module: Optional[Module] = None,
     ):
-        super().__init__(out_channels, stats_list)
+        super().__init__(out_channels, stats_list, post_module)
 
     @abstractmethod
     def forward(self, x: Tensor):
@@ -37,7 +49,27 @@ class StypeEncoder(Module, ABC):
 
     @abstractmethod
     def reset_parameters(self):
-        raise NotImplementedError
+        r"""Initialize the parameters of `post_module`"""
+        if self.post_module is not None:
+            if isinstance(self.post_module, Sequential):
+                for m in self.post_module:
+                    reset_parameters_soft(m)
+            else:
+                reset_parameters_soft(self.post_module)
+
+    def post_forward(self, out: Tensor) -> Tensor:
+        r"""Post-forward function applied to :obj:`out` of shape
+        [batch_size, num_cols, channels]. It also returns :obj:`out` of the
+        same shape."""
+        if self.post_module is not None:
+            shape_before = out.shape
+            out = self.post_module(out)
+            if out.shape != shape_before:
+                raise RuntimeError(
+                    f"post_module must not alter the shape of the tensor, but "
+                    f"it changed the shape from {shape_before} to "
+                    f"{out.shape}.")
+        return out
 
 
 class EmbeddingEncoder(StypeEncoder):
@@ -50,8 +82,9 @@ class EmbeddingEncoder(StypeEncoder):
         self,
         out_channels: Optional[int] = None,
         stats_list: Optional[List[Dict[StatType, Any]]] = None,
+        post_module: Optional[Module] = None,
     ):
-        super().__init__(out_channels, stats_list)
+        super().__init__(out_channels, stats_list, post_module)
 
     def init_modules(self):
         self.embs = ModuleList([])
@@ -65,6 +98,11 @@ class EmbeddingEncoder(StypeEncoder):
                     padding_idx=0,
                 ))
         self.reset_parameters()
+
+    def reset_parameters(self):
+        super().reset_parameters()
+        for emb in self.embs:
+            emb.reset_parameters()
 
     def forward(self, x: Tensor):
         r"""Maps input :obj:`x` from TensorFrame (shape [batch_size, num_cols])
@@ -82,12 +120,8 @@ class EmbeddingEncoder(StypeEncoder):
         for i, emb in enumerate(self.embs):
             xs.append(emb(x[:, i]))
         # [batch_size, num_cols, hidden_channels]
-        x = torch.stack(xs, dim=1)
-        return x
-
-    def reset_parameters(self):
-        for emb in self.embs:
-            emb.reset_parameters()
+        out = torch.stack(xs, dim=1)
+        return self.post_forward(out)
 
 
 class LinearEncoder(StypeEncoder):
@@ -101,8 +135,9 @@ class LinearEncoder(StypeEncoder):
         self,
         out_channels: Optional[int] = None,
         stats_list: Optional[List[Dict[StatType, Any]]] = None,
+        post_module: Optional[Module] = None,
     ):
-        super().__init__(out_channels, stats_list)
+        super().__init__(out_channels, stats_list, post_module)
 
     def init_modules(self):
         mean = torch.tensor(
@@ -115,6 +150,11 @@ class LinearEncoder(StypeEncoder):
         self.weight = Parameter(torch.empty(num_cols, self.out_channels))
         self.bias = Parameter(torch.empty(num_cols, self.out_channels))
         self.reset_parameters()
+
+    def reset_parameters(self):
+        super().reset_parameters()
+        torch.nn.init.normal_(self.weight, std=0.1)
+        torch.nn.init.zeros_(self.bias)
 
     def forward(self, x: Tensor):
         r"""Maps input :obj:`x` from TensorFrame (shape [batch_size, num_cols])
@@ -129,11 +169,8 @@ class LinearEncoder(StypeEncoder):
         # [batch_size, num_cols, channels] + [num_cols, channels]
         # -> [batch_size, num_cols, channels]
         x = x_lin + self.bias
-        return torch.nan_to_num(x, nan=0)
-
-    def reset_parameters(self):
-        torch.nn.init.normal_(self.weight, std=0.1)
-        torch.nn.init.zeros_(self.bias)
+        out = torch.nan_to_num(x, nan=0)
+        return self.post_forward(out)
 
 
 class LinearBucketEncoder(StypeEncoder):
@@ -146,8 +183,9 @@ class LinearBucketEncoder(StypeEncoder):
         self,
         out_channels: Optional[int] = None,
         stats_list: Optional[List[Dict[StatType, Any]]] = None,
+        post_module: Optional[Module] = None,
     ):
-        super().__init__(out_channels, stats_list)
+        super().__init__(out_channels, stats_list, post_module)
 
     def init_modules(self):
         # The min, 25th, 50th, 75th quantile, and max of the column.
@@ -159,6 +197,12 @@ class LinearBucketEncoder(StypeEncoder):
             torch.empty(num_cols, self.interval.shape[-1], self.out_channels))
         self.bias = Parameter(torch.empty(num_cols, self.out_channels))
         self.reset_parameters()
+
+    def reset_parameters(self):
+        super().reset_parameters()
+        # Reset learnable parameters of the linear transformation
+        torch.nn.init.normal_(self.weight, std=0.1)
+        torch.nn.init.zeros_(self.bias)
 
     def forward(self, x: Tensor):
         encoded_values = []
@@ -188,12 +232,8 @@ class LinearBucketEncoder(StypeEncoder):
         # -> [batch_size, num_cols, channels]
         x_lin = torch.einsum('ijk,jkl->ijl', out, self.weight)
         x = x_lin + self.bias
-        return torch.nan_to_num(x, nan=0)
-
-    def reset_parameters(self):
-        # Reset learnable parameters of the linear transformation
-        torch.nn.init.normal_(self.weight, std=0.1)
-        torch.nn.init.zeros_(self.bias)
+        out = torch.nan_to_num(x, nan=0)
+        return self.post_forward(out)
 
 
 class LinearPeriodicEncoder(StypeEncoder):
@@ -212,10 +252,11 @@ class LinearPeriodicEncoder(StypeEncoder):
         self,
         out_channels: Optional[int] = None,
         stats_list: Optional[List[Dict[StatType, Any]]] = None,
+        post_module: Optional[Module] = None,
         n_bins: Optional[int] = 16,
     ):
         self.n_bins = n_bins
-        super().__init__(out_channels, stats_list)
+        super().__init__(out_channels, stats_list, post_module)
 
     def init_modules(self):
         mean = torch.tensor(
@@ -230,6 +271,11 @@ class LinearPeriodicEncoder(StypeEncoder):
             torch.empty((num_cols, self.n_bins * 2, self.out_channels)))
         self.reset_parameters()
 
+    def reset_parameters(self):
+        super().reset_parameters()
+        torch.nn.init.normal_(self.linear_in, std=0.1)
+        torch.nn.init.normal_(self.linear_out, std=0.1)
+
     def forward(self, x: Tensor):
         x = (x - self.mean) / self.std
         # Compute the value 'v' by scaling the input 'x' with
@@ -243,9 +289,5 @@ class LinearPeriodicEncoder(StypeEncoder):
         # [batch_size, num_cols, num_buckets],[num_cols, num_buckets, channels]
         # -> [batch_size, num_cols, channels]
         x = torch.einsum('ijk,jkl->ijl', x, self.linear_out)
-
-        return torch.nan_to_num(x, nan=0)
-
-    def reset_parameters(self):
-        torch.nn.init.normal_(self.linear_in, std=0.1)
-        torch.nn.init.normal_(self.linear_out, std=0.1)
+        out = torch.nan_to_num(x, nan=0)
+        return self.post_forward(out)